+++ src/Symfony/Bridge/Doctrine/DataCollector/DoctrineDataCollector.php
@@ -164,6 +164,8 @@ class DoctrineDataCollector extends DataCollector
             }
         }
 
+        $query['params'] = $this->cloneVar($query['params']);
+
         return $query;
     }