+++ src/Symfony/Component/DependencyInjection/Loader/Configurator/ServicesConfigurator.php
@@ -139,4 +139,9 @@ class ServicesConfigurator extends AbstractConfigurator
     {
         return $this->set($id, $class);
     }
+
+    public function __destruct()
+    {
+        $this->loader->registerAliasesForSinglyImplementedInterfaces();
+    }
 }