+++ src/Symfony/Component/Messenger/DependencyInjection/MessengerPass.php
@@ -70,6 +70,7 @@ class MessengerPass implements CompilerPassInterface
     {
         $definitions = [];
         $handlersByBusAndMessage = [];
+        $handlerToOriginalServiceIdMapping = [];
 
         foreach ($container->findTaggedServiceIds($this->handlerTag, true) as $serviceId => $tags) {
             foreach ($tags as $tag) {
@@ -109,6 +110,10 @@ class MessengerPass implements CompilerPassInterface
                         $options = ['method' => $options];
                     }
 
+                    if (!isset($options['from_transport']) && isset($tag['from_transport'])) {
+                        $options['from_transport'] = $tag['from_transport'];
+                    }
+
                     $priority = $tag['priority'] ?? $options['priority'] ?? 0;
                     $method = $options['method'] ?? '__invoke';
 
@@ -139,6 +144,8 @@ class MessengerPass implements CompilerPassInterface
                     } else {
                         $definitionId = $serviceId;
                     }
+
+                    $handlerToOriginalServiceIdMapping[$definitionId] = $serviceId;
 
                     foreach ($buses as $handlerBus) {
                         $handlersByBusAndMessage[$handlerBus][$message][$priority][] = [$definitionId, $options];
@@ -188,6 +195,12 @@ class MessengerPass implements CompilerPassInterface
             foreach ($busIds as $bus) {
                 if (!isset($debugCommandMapping[$bus])) {
                     $debugCommandMapping[$bus] = [];
+                }
+
+                foreach ($debugCommandMapping[$bus] as $message => $handlers) {
+                    foreach ($handlers as $key => $handler) {
+                        $debugCommandMapping[$bus][$message][$key][0] = $handlerToOriginalServiceIdMapping[$handler[0]];
+                    }
                 }
             }
             $container->getDefinition('console.command.messenger_debug')->replaceArgument(0, $debugCommandMapping);